-Added a template to help with the creation of the TemplateContextMenus.cs file. There is also a smaller template inside the TemplateEditor that could be abstracted out as well.
Added an Scriptable Object called TemplateEditor and a custom inspector for it to allow for easy management.

Working version of the ContextMenu creator, just by editing the TemplateEditor you can now generator the context menus for an array of templates as well as define their category and priority in the context menu.